requirements.txt
# Core dependencies
playwright==1.40.0
Pillow==10.1.0
pillow-heif>=0.13.0  # HEIC/HEIF image support
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1

# Vision AI Backends
# Claude (Anthropic)
anthropic>=0.40.0

# BLIP-2 (Local - FREE, no API key needed)
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0  # For efficient model loading

# OpenAI GPT-4 Vision (optional)
openai>=1.3.0

# Google Gemini Vision (optional)
google-generativeai>=0.3.0

# Required for compatibility
httpx>=0.27.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
